Docker/requirements.txt
pandas==2.1.4
numpy==1.26.2
psycopg2-binary==2.9.9
